feat(api-headless-cms-ddb-es): filtering of root level query
