+++ packages/api-headless-cms-ddb-es/src/elasticsearch/search/refSearch.ts
@@ -4,11 +4,6 @@ import {
     TransformCallable
 } from "~/plugins/CmsEntryElasticsearchQueryBuilderValueSearchPlugin";
 
-interface RefValue {
-    id?: string;
-    entryId: string;
-}
-
 const createPath: CreatePathCallable<string> = ({ field, key }) => {
     if (key && key.match("entryId") === null) {
         return `${field.storageId}.id`;
@@ -16,14 +11,8 @@ const createPath: CreatePathCallable<string> = ({ field, key }) => {
     return `${field.storageId}.entryId`;
 };
 
-const transform: TransformCallable<RefValue | string> = ({ value }) => {
-    if (typeof value === "string") {
-        return value;
-    }
-    if (value.id) {
-        return value.id;
-    }
-    return value.entryId;
+const transform: TransformCallable = ({ value }) => {
+    return value;
 };
 
 export const createRefSearchPlugin = (): CmsEntryElasticsearchQueryBuilderValueSearchPlugin => {
