+++ packages/api-headless-cms-ddb-es/src/operations/entry/elasticsearch/filtering/populated.ts
@@ -1,0 +1,20 @@
+import { ElasticsearchBoolQueryConfig } from "@webiny/api-elasticsearch/types";
+
+export const getPopulated = (
+    query: ElasticsearchBoolQueryConfig
+): Partial<ElasticsearchBoolQueryConfig> => {
+    const result: Partial<ElasticsearchBoolQueryConfig> = {};
+    let key: keyof ElasticsearchBoolQueryConfig;
+    for (key in query) {
+        const value = query[key];
+        if (value === undefined || (Array.isArray(value) && value.length === 0)) {
+            continue;
+        }
+        /**
+         * TODO figure out better types.
+         */
+        // @ts-ignore
+        result[key] = value;
+    }
+    return result;
+};
